==> arxiv_data/step_1.ipynb <==
#マウント
from google.colab import drive
drive.mount('/content/drive')

##作業するディレクトリを指定する.
%cd '/content/drive/MyDrive/Colab_Notebooks/arXiv_data'

#関係するツールの導入

#pdfをtxtファイルに変換する際に使用
from bs4 import BeautifulSoup
import urllib
import urllib.request as req

import time
import requests
import os
import numpy as np
import re

#input_folder_path: 入力となるファイルの保存場所
input_folder_path ='/content/drive/MyDrive/Colab_Notebooks/arXiv_data/Step_1'
#output_folder_path: 出力となるpdfファイルの保存場所
output_folder_path = '/content/drive/MyDrive/Colab_Notebooks/arXiv_data/Step_2'

#PDFをダウンロードするためにpdf2txtをインストールする．
!pip install PyPDF2

!pip install pdfminer.six

#pdf2txt.pyのダウンロード
py_url ='https://github.com/pdfminer/pdfminer.six/blob/master/tools/pdf2txt.py'
py_fn='pdf2txt.py'
os.system("wget -O " + str(py_fn) + " " + str(py_url))

module_path = '/content/drive/My Drive/Colab_Notebooks/arXiv_data/module/'
os.chdir(module_path)
!ls

#ファイル名の一覧を作る．
import file_processing

file_names_list = file_processing.read_file_names_list( input_folder_path, '.pdf' , f = lambda x : os.path.splitext(x)[0] )

file_names_list

len(file_names_list)

#get_txt_from_pdf (str,str,str)->
#(入力の場所,出力の場所,ファイルの名前)->txtファイルを出力の場所に作る. 
#
#
#('./folder_1', './folder_2', file_name )-> ['I'm a doctor.', 'My ... .']
#
# get_txt_from_pdf (import_folder_path,export_folder_path,file_name)
#
#param string input_file_path 読み込みたいファイルの入っているフォルダへのpath
#param string export_file_path 書き込んだファイルを保存するファイルへのpath
#param string file_name 実行対象となるファイルの名前
#
#return なし
#       export_folder_path に file_name + .txt で保存する．
#       import_pdfpath  のファイルを消去する． 
 
def get_txt_from_pdf (import_folder_path,export_folder_path,file_name):
  import_pdfpath = import_folder_path + "/" + file_name + ".pdf"
  export_pdfpath = export_folder_path + "/" 
  #pdf2txt.pyを用いてpdfファイルをテキスト変換
  lines = !pdf2txt.py { import_pdfpath }
  txt = '\n'.join(lines)
  rename = export_pdfpath + file_name + ".txt"
  file_processing.write_file_from_list( rename , txt)
  os.remove( import_pdfpath )

#get_txt_from_pdfをlistに対して繰り返し実行
list( map ( lambda x: get_txt_from_pdf(input_folder_path, output_folder_path, x), file_names_list))
