==> arxiv_data/step_2.ipynb <==
#マウント
from google.colab import drive
drive.mount('/content/drive')

##作業するディレクトリを指定する.
%cd '/content/drive/MyDrive/Colab_Notebooks/test'

#関係するツールの導入
import os
import numpy as np
import pandas as pd
import collections
import re
from string import digits

#input_folder_path: 入力となるファイルの保存場所
input_folder_path = '/content/drive/MyDrive/Colab_Notebooks/test/txt_file'
#output_folder_path: 出力となるpdfファイルの保存場所
output_folder_path = '/content/drive/MyDrive/Colab_Notebooks/test/step3'

#モジュール
module_path = '/content/drive/My Drive/Colab_Notebooks/arXiv_data/module/'
os.chdir(module_path)
import file_processing

file_names_list = file_processing.read_file_names_list( input_folder_path, '.txt' , f = lambda x : os.path.splitext(x)[0] )

file_names_list

len(file_names_list)

file_path = input_folder_path + "/" + file_names_list[0] + ".txt"
with open(file_path) as f:
  r = f.read()
  print(r)

#take_out_word (str, str)-> str_lst
#(入力の場所,ファイルの名前)->txtファイルを読み込んで単語に分ける.小文字にして,ピリオドなどは除く. 
#doxygen文法
#
#指定テキストファイル内の英文を単語ごとの配列に整形します。
#主に半角スペース、コンマ、ピリオド、ハイフンで単語を分割します。
#また、数字の除去と大文字の小文字化も行っています。
#出力は文字列のリスト型で、順番は登場順になります。
#
#例
#"I'm a apple-pen."->["i","m","a","apple","pen"]
#
#param string path 読み込みファイルが存在するフォルダパス
#param string file_name 読み込むファイルの名前(拡張子込み)
#
#return string[] 分割された単語のリスト
#
#split_txt_to_words_list(read_folder_path,read_file_name)
def take_out_word (path,file_name):
  # split()でスペースと改行で分割したリストから単語を取り出す
  def word_list_of(string_file):
    #小文字にしてから英単語
    table = str.maketrans('', '', digits)
    string_file = string_file.translate(table)
    string_file =string_file.lower()
    string_beta = re.findall("\w+",string_file)
    return(string_beta)
  file_path = path + "/" + file_name + ".txt"
  with open(file_path) as f:
    r = f.read()
    s = word_list_of(r)
  return s

#get_lst_of_word (str,str,str)->
#(入力の場所,出力の場所,ファイルの名前)->take_out_wordを実行して出力の場所にtxtファイルで保存する.
#
#~~~
#内部でtext_out_word()を用いています
#
#
def get_lst_of_word (input_folder_path, output_folder_path, file_name):
  ##pdfpath_1 = path_1 + "/" + file_name
  lst = take_out_word(input_folder_path, file_name)
  output_file_path = output_folder_path + "/" + file_name + ".txt"
  file_processing.write_file_from_list(output_file_path , lst)
  os.remove(input_folder_path + "/" + file_name + ".txt" )

#get_lst_of_wordをlstに対して繰り返し実行する.
txt_data =list(map( lambda x : get_lst_of_word(input_folder_path, output_folder_path, x), file_names_list))
