==> arxiv_data/step_3.ipynb <==
#マウント
from google.colab import drive
drive.mount('/content/drive')
%cd "drive/MyDrive"

##作業するディレクトリを指定する.
%cd '/content/drive/MyDrive/Colab_Notebooks/arXiv_data'

#関係するツールの導入
import os
import numpy as np
import pandas as pd
import collections
import re
from string import digits

#input_folder_path: 入力となるファイルの保存場所
input_folder_path ='/content/drive/MyDrive/Colab_Notebooks/arXiv_data/Step_3'
#output_folder_path: 出力となるpdfファイルの保存場所
output_folder_path = '/content/drive/MyDrive/Colab_Notebooks/arXiv_data/Step_4'

#モジュール
module_path = '/content/drive/My Drive/Colab_Notebooks/arXiv_data/module/'
os.chdir(module_path)
import file_processing

file_names_list = file_processing.read_file_names_list( input_folder_path, '.txt' , f = lambda x : os.path.splitext(x)[0] )

len(file_names_list)

from pandas.core.frame import DataFrame
#read_lst (str,str)->dct
#(入力の場所,ファイルの名前)->単語を数えてdct型にする(key,value)->(str,int)
def read_lst (input_folder_path, output_folder_path, file_name):
  file_path = input_folder_path + "/" + file_name + ".txt"
  list_row = file_processing.read_file_to_list(file_path)
  word_lst = collections.Counter(list_row)
  df = pd.DataFrame( word_lst.values(), index = word_lst.keys()).T
  df.to_csv(output_folder_path + '/' + file_name + '.csv', sep=",")
  os.remove(file_path)

#read_lstをlstに対して繰り返し実行する.
txt_data =list(map( lambda x : read_lst(input_folder_path, output_folder_path, x), file_names_list))
