==> arxiv_data/step_4.ipynb <==
#マウント
from google.colab import drive
drive.mount('/content/drive')
%cd "drive/MyDrive"

##作業するディレクトリを指定する.
%cd '/content/drive/MyDrive/Colab_Notebooks/arXiv_data'

#関係するツールの導入
import os
import numpy as np
import pandas as pd
import collections
import re
from string import digits

import sys
import time

#input_folder_path: 入力となるファイルの保存場所
input_folder_path ='/content/drive/MyDrive/Colab_Notebooks/arXiv_data/Step_4'
#output_folder_path: 出力となるpdfファイルの保存場所
output_folder_path = '/content/drive/MyDrive/Colab_Notebooks/arXiv_data/Step_5'

#モジュール
module_path = '/content/drive/My Drive/Colab_Notebooks/arXiv_data/module/'
os.chdir(module_path)
import file_processing

file_names_list = file_processing.read_file_names_list( input_folder_path, '.csv' , f = lambda x : os.path.splitext(x)[0] )
file_names_list.sort(reverse=False)

file_processing.write_file_from_list(output_folder_path + "/" + '/file_names_list_10000.txt', file_names_list)

file_names_list = file_processing.read_file_to_list(output_folder_path + "/" + 'file_names_list_10000.txt')

len(file_names_list)

#すべてのファイルで出現する単語のリストを作る．
start = time.perf_counter()
df_dict = []
for i in range(len(file_names_list)):
  input_file_path = input_folder_path + '/' + file_names_list[i] + '.csv'
  d = pd.read_csv(input_file_path)
  d = d.drop('Unnamed: 0', axis=1)
  d = list(d.columns.values)
  df_dict = df_dict+ d
df_dict = list(set(df_dict))
print(time.perf_counter() - start)
print(sys.getsizeof(df_dict)/(1024**2))

print(len(df_dict))

file_processing.write_file_from_list(output_folder_path + "/" + 'df_dict.txt', df_dict)

df_dict = file_processing.read_file_to_list(output_folder_path + "/" + 'df_dict.txt')

#出現する単語のリストを3文字以上の単語のみに減らす．
df_dict_3 = [ l for l in df_dict if  len( l ) > 3]

file_processing.write_file_from_list(output_folder_path + "/" + 'df_dict_3.txt', df_dict_3 )

#すべてのファイルで出現する単語数を数える．
start = time.perf_counter()

file_names_list_number = len( file_names_list)
ar = np.zeros((2, len(df_dict_3) ))
df = pd.DataFrame(ar, columns = df_dict_3 )
df.insert(0, '全単語数',df.sum(axis=1))

for i in range( file_names_list_number ):
  input_file_path = input_folder_path + '/' + file_names_list[i] + '.csv'
  d = pd.read_csv(input_file_path)
  d = d.drop('Unnamed: 0', axis=1)
  d.insert(0, '全単語数',d.sum(axis=1))
  df.iloc[0, : ] = d.iloc[0,:]
  df = df.fillna(0)
  sum_df = df.loc[0] + df.loc[1]
  df.iloc[1, : ] = sum_df

print(time.perf_counter() - start)
print(sys.getsizeof(df)/(1024**2))

(df.loc[[1]]).to_csv(output_folder_path + '/' +"df_dict_data_10000.csv", sep=",")
